Fix and reformat

+++ mlbench_core/evaluation/pytorch/criterion.py
@@ -168,7 +168,7 @@ class LabelSmoothing(nn.Module):
         self.smoothing = smoothing
 
     def forward(self, x, target):
-        logprobs = torch.nn.functional.log_softmax(x, dim=-1)
+        logprobs = torch.nn.functional.log_softmax(x, dim=-1).type(torch.float32)
 
         non_pad_mask = target != self.padding_idx
         nll_loss = -logprobs.gather(dim=-1, index=target.unsqueeze(1))
